+++ tests/test_examples.py
@@ -2,7 +2,7 @@ from unittest import TestCase
 import os
 from os.path import join
 
-from soil import utils, simulation
+from soil import serialization, simulation
 
 ROOT = os.path.abspath(os.path.dirname(__file__))
 EXAMPLES = join(ROOT, '..', 'examples')
@@ -17,31 +17,32 @@ class TestExamples(TestCase):
 def make_example_test(path, config):
     def wrapped(self):
         root = os.getcwd()
-        os.chdir(os.path.dirname(path))
-        s = simulation.from_config(config)
-        iterations = s.max_time * s.num_trials
-        if iterations > 1000:
-            s.max_time = 100
-            s.num_trials = 1
-        if config.get('skip_test', False) and not FORCE_TESTS:
-            self.skipTest('Example ignored.')
-        envs = s.run_simulation(dry_run=True)
-        assert envs
-        for env in envs:
-            assert env
-            try:
-                n = config['network_params']['n']
-                assert len(list(env.network_agents)) == n
-                assert env.now > 2  # It has run
-                assert env.now <= config['max_time']  # But not further than allowed
-            except KeyError:
-                pass
-        os.chdir(root)
+        for s in simulation.all_from_config(path):
+            iterations = s.max_time * s.num_trials
+            if iterations > 1000:
+                s.max_time = 100
+                s.num_trials = 1
+            if config.get('skip_test', False) and not FORCE_TESTS:
+                self.skipTest('Example ignored.')
+            envs = s.run_simulation(dry_run=True)
+            assert envs
+            for env in envs:
+                assert env
+                try:
+                    n = config['network_params']['n']
+                    assert len(list(env.network_agents)) == n
+                    assert env.now > 2  # It has run
+                    assert env.now <= config['max_time']  # But not further than allowed
+                except KeyError:
+                    pass
     return wrapped
 
 
 def add_example_tests():
-    for config, path in utils.load_config(join(EXAMPLES, '**', '*.yml')):
+    for config, path in serialization.load_files(
+            join(EXAMPLES, '*', '*.yml'),
+            join(EXAMPLES, '*.yml'),
+    ):
         p = make_example_test(path=path, config=config)
         fname = os.path.basename(path)
         p.__name__ = 'test_example_file_%s' % fname