+++ CHANGELOG.md
@@ -3,6 +3,19 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.14.0]
+### Added
+* Loading configuration from template definitions in the yaml, in preparation for SALib support.
+The definition of the variables and their possible values (i.e., a problem in SALib terms), as well as a sampler function, can be provided.
+Soil uses this definition and the template to generate a set of configurations.
+* Simulation group names, to link related simulations. For now, they are only used to group all simulations in the same group under the same folder.
+
+### Changed
+* `dir_path` is now the directory for resources (modules, files), and the output dir has been renamed to `outdir`
+
+## Fixed
+* Modules with custom agents/environments failed to load when they were run from outside the directory of the definition file. Modules are now loaded from the directory of the simulation file in addition to the working directory
+
 ## [0.13.8]
 ### Changed
 * Moved TerroristNetworkModel to examples